grpc-demo/pyproject.toml
[tool.poetry]
name = "grpc-demo"
version = "0.1.0"
description = ""
authors = ["Mathieu Lambert"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.64.0"
grpcio-tools = "^1.64.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
